Update JobSchemaV3 to JobSchemaOld for compatibility
